% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/airport_codes.R
\name{airport_codes_by_city}
\alias{airport_codes_by_city}
\title{Get airport codes by city}
\usage{
airport_codes_by_city(the_city, the_state = "")
}
\arguments{
\item{the_city}{string, name of US city}

\item{the_state}{string, optional, two-letter code for US state}
}
\value{
data frame with Code, City, State, and Airport
}
\description{
Get airport codes by city
}
\examples{
airport_codes_by_city("New York")
airport_codes_by_city("Greenville", "NC")
}
