% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/airport_codes.R
\name{airport_codes_by_name}
\alias{airport_codes_by_name}
\title{Get airport code by airport name}
\usage{
airport_codes_by_name(the_airport)
}
\arguments{
\item{the_airport}{string, name or partial name of airport}
}
\value{
data frame with Code, City, State, and Airport
}
\description{
Get airport code by airport name
}
\examples{
airport_codes_by_name("Kennedy")

}
