% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/airport_codes.R
\name{airport_codes_by_state}
\alias{airport_codes_by_state}
\title{Get all airport codes in a state}
\usage{
airport_codes_by_state(the_state)
}
\arguments{
\item{the_state}{string, two-letter code for a US state}
}
\value{
data frame with Code, City, State, and Airport
}
\description{
Get all airport codes in a state
}
\examples{
airport_codes_by_state("RI")
}
